middleware/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.13
pydantic[email]==1.10.13
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
requests==2.31.0
httpx==0.24.1
aiofiles==23.2.1
jinja2==3.1.2
python-dotenv==1.0.0
supabase==2.0.0
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
openai==1.54.3
geopy==2.4.1
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2
matplotlib==3.7.5
seaborn==0.13.0
plotly==5.17.0
websockets==12.0
email-validator==2.1.0
gunicorn==21.2.0 